+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
 ## Unreleased changes
-
 - Added support for GRPC V2 Get Account Info
 - Added support for GRPC V2 Get Account List
 - Added support for GRPC V2 Get Block Items
+- Added support for GRPC V2 Get Consensus Status/Info
 
 ## 4.2.0
 - Added initial support for GRPC V2